requirements.txt
Django==5.1.3 ; python_version >= "3.11"
autopep8==2.3.1
django-compressor==4.5.1
django-environ==0.11.2
django-oauth-toolkit==3.0.1
django-sass-processor==1.4.1
django-structlog==9.1.1
django-tree-queries==0.19.0
djangorestframework==3.15.2
factory_boy==3.3.1
html-sanitizer==2.4.4
libsass==0.23.0
Markdown==3.7
pip-tools==7.4.1
pre_commit==4.2.0
psycopg2==2.9.10
pytest==8.3.3
pytest-cov==6.0.0
pytest-django==4.9.0
requests==2.32.3
rich==14.0.0
ruff==0.11.2
structlog==25.2.0
httpie==3.2.4